pyproject.toml
[tool.poetry]
name = "eda_frequency_api"
version = "0.1.0"
description = ""
authors = ["Jonas Stettner <mail@jstet.net>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
databases = "^0.6.1"
python-dotenv = "^0.20.0"
asyncpg = "^0.26.0"
psycopg2 = "^2.9.3"
sqlacodegen = "^2.3.0"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
start = "eda_frequency_api.main:start"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
